Iter3 IT12 IT15 IT17 IT19 IT20 IT33 IT34 MQ7 MQ8x2 MQ9 MQ10 MQ11 MQ12

+++ la-caja/libraries/EspInfInd/EspInfInd.h
@@ -35,7 +35,7 @@ class EspInfInd
     void Loop();
     void MqttReceived(char* strTopic, byte* payload, unsigned int length);
     void MqttSend(char* strTopic, char* strGameStatus, const char *strSrc=STR_ORG_BOARD);
-    void UpdateSwitch(int iUpdateType, long newLevel, long newConfig);
+    void UpdateSwitch(int iUpdateType, int newLevel, int newConfig);
     ~EspInfInd();
     
     // REQ.MQ1
